Internal: identity user cleanup
